requirements.txt
streamlit==1.26.1
pdfplumber==0.7.6
python-docx==0.8.11
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.25.2